third_party/blink/web_tests/fast/js/dfg-peephole-compare-final-object-to-final-object-or-other-expected.txt

Tests that (Peephole) CompareEq for the case where one side is predicted final object and the other side is predicted either final object or other (i.e. null or undefined) when document.all is present.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS foo({f:42}, null) is false
PASS foo(o, o) is true
PASS foo({f:42}, null) is false
PASS foo(o, o) is true
PASS foo({f:42}, null) is false
PASS foo(o, o) is true
PASS foo({f:42}, null) is false
PASS foo(o, o) is true
PASS foo({f:42}, null) is false
PASS foo(o, o) is true
PASS foo({f:42}, null) is false
PASS foo(o, o) is true
PASS foo({f:42}, null) is false
PASS foo(o, o) is true
PASS foo({f:42}, null) is false
PASS foo(o, o) is true
PASS foo({f:42}, null) is false
PASS foo(o, o) is true
PASS foo({f:42}, null) is false
PASS foo(o, o) is true
PASS foo({f:42}, null) is false
PASS foo(o, o) is true
PASS foo({f:42}, null) is false
PASS foo(o, o) is true
PASS foo({f:42}, null) is false
PASS foo(o, o) is true
PASS foo({f:42}, null) is false
PASS foo(o, o) is true
PASS foo({f:42}, null) is false
PASS foo(o, o) is true
PASS foo({f:42}, null) is false
PASS foo(o, o) is true
PASS foo({f:42}, null) is false
PASS foo(o, o) is true
PASS foo({f:42}, null) is false
PASS foo(o, o) is true
PASS foo({f:42}, null) is false
PASS foo(o, o) is true
PASS foo({f:42}, null) is false
PASS foo(o, o) is true
PASS foo({f:42}, null) is false
PASS foo(o, o) is true
PASS foo({f:42}, null) is false
PASS foo(o, o) is true
PASS foo({f:42}, null) is false
PASS foo(o, o) is true
PASS foo({f:42}, null) is false
PASS foo(o, o) is true
PASS foo({f:42}, null) is false
PASS foo(o, o) is true
PASS foo({f:42}, null) is false
PASS foo(o, o) is true
PASS foo({f:42}, null) is false
PASS foo(o, o) is true
PASS foo({f:42}, null) is false
PASS foo(o, o) is true
PASS foo({f:42}, null) is false
PASS foo(o, o) is true
PASS foo({f:42}, null) is false
PASS foo(o, o) is true
PASS foo({f:42}, null) is false
PASS foo(o, o) is true
PASS foo({f:42}, null) is false
PASS foo(o, o) is true
PASS foo({f:42}, null) is false
PASS foo(o, o) is true
PASS foo({f:42}, null) is false
PASS foo(o, o) is true
PASS foo({f:42}, null) is false
PASS foo(o, o) is true
PASS foo({f:42}, null) is false
PASS foo(o, o) is true
PASS foo({f:42}, null) is false
PASS foo(o, o) is true
PASS foo({f:42}, null) is false
PASS foo(o, o) is true
PASS foo({f:42}, null) is false
PASS foo(o, o) is true
PASS foo({f:42}, null) is false
PASS foo(o, o) is true
PASS foo({f:42}, null) is false
PASS foo(o, o) is true
PASS foo({f:42}, null) is false
PASS foo(o, o) is true
PASS foo({f:42}, null) is false
PASS foo(o, o) is true
PASS foo({f:42}, null) is false
PASS foo(o, o) is true
PASS foo({f:42}, null) is false
PASS foo(o, o) is true
PASS foo({f:42}, null) is false
PASS foo(o, o) is true
PASS foo({f:42}, null) is false
PASS foo(o, o) is true
PASS foo({f:42}, null) is false
PASS foo(o, o) is true
PASS foo({f:42}, null) is false
PASS foo(o, o) is true
PASS foo({f:42}, null) is false
PASS foo(o, o) is true
PASS successfullyParsed is true

TEST COMPLETE
